#4: Rename get_api_endpoint - Renamed to improve clarity of component
